requirements.txt
google-generativeai
praw
pandas